Receiver Token
Receiver token now creates a token.txt the first time it receives data. If it the file already exists it will compare the data in the file to token value sent by sender, if token is in database it will respond with "TOKEN IS IN DATABASE", if it is not there it is meant to add it as a new connection into token.txt, alongside the IP address and port number.

I have tested out the system with another unique token value. It just replaces the first line of the token file. which is something i plan on fixing tomorrow.

I couldn't get the receivers Ip address which is why it is not stored in the same token.txt, which would have then messed up the formatting of the token.txt file, but a separate token file called mytoken.txt. It may not need to be fixed but idk right now.